Reorganize code so crypto and utils are in sub-packages.
