+++ src/main/java/org/jitsi/srtp/SrtpCryptoContext.java
@@ -38,6 +38,7 @@ import java.util.*;
 
 import org.bouncycastle.crypto.params.*;
 import org.jitsi.bccontrib.params.*;
+import org.jitsi.srtp.utils.*;
 import org.jitsi.utils.*;
 import org.jitsi.utils.logging.*;
