+++ src/main/java/org/jitsi/srtp/crypto/OpenSslWrapperLoader.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.jitsi.srtp;
+package org.jitsi.srtp.crypto;
 
 import org.jitsi.utils.*;
 import org.jitsi.utils.logging.*;
