+++ src/main/java/org/jitsi/srtp/crypto/SrtpCipherCtrJava.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.srtp;
+package org.jitsi.srtp.crypto;
 
 import org.bouncycastle.crypto.*;
 import org.bouncycastle.crypto.params.*;
