+++ src/main/java/org/jitsi/srtp/utils/SrtpPacket.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jitsi.srtp;
+package org.jitsi.srtp.utils;
 
 import org.jitsi.utils.*;
