+++ src/main/native/Hmac.c
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "org_jitsi_srtp_OpenSslHmac.h"
+#include "org_jitsi_srtp_crypto_OpenSslHmac.h"
 
 #include <openssl/evp.h>
 #include <openssl/hmac.h>
@@ -27,7 +27,7 @@
  * Signature: (J)I
  */
 JNIEXPORT jint JNICALL
-Java_org_jitsi_srtp_OpenSslHmac_EVP_1MD_1size
+Java_org_jitsi_srtp_crypto_OpenSslHmac_EVP_1MD_1size
     (JNIEnv *env, jclass clazz, jlong md)
 {
     return EVP_MD_size((const EVP_MD *) (intptr_t) md);
@@ -39,7 +39,7 @@ Java_org_jitsi_srtp_OpenSslHmac_EVP_1MD_1size
  * Signature: ()J
  */
 JNIEXPORT jlong JNICALL
-Java_org_jitsi_srtp_OpenSslHmac_EVP_1sha1
+Java_org_jitsi_srtp_crypto_OpenSslHmac_EVP_1sha1
     (JNIEnv *env, jclass clazz)
 {
     return (jlong) (intptr_t) EVP_sha1();
@@ -51,7 +51,7 @@ Java_org_jitsi_srtp_OpenSslHmac_EVP_1sha1
  * Signature: ()J
  */
 JNIEXPORT jlong JNICALL
-Java_org_jitsi_srtp_OpenSslHmac_HMAC_1CTX_1create
+Java_org_jitsi_srtp_crypto_OpenSslHmac_HMAC_1CTX_1create
     (JNIEnv *env, jclass clazz)
 {
 /* OpenSSL 1.1.0 made HMAC_CTX an opaque structure, which must be allocated
@@ -76,7 +76,7 @@ Java_org_jitsi_srtp_OpenSslHmac_HMAC_1CTX_1create
  * Signature: (J)V
  */
 JNIEXPORT void JNICALL
-Java_org_jitsi_srtp_OpenSslHmac_HMAC_1CTX_1destroy
+Java_org_jitsi_srtp_crypto_OpenSslHmac_HMAC_1CTX_1destroy
     (JNIEnv *env, jclass clazz, jlong ctx)
 {
     HMAC_CTX *ctx_ = (HMAC_CTX *) (intptr_t) ctx;
@@ -97,7 +97,7 @@ Java_org_jitsi_srtp_OpenSslHmac_HMAC_1CTX_1destroy
  * Signature: (J[BII)I
  */
 JNIEXPORT jint JNICALL
-Java_org_jitsi_srtp_OpenSslHmac_HMAC_1Final
+Java_org_jitsi_srtp_crypto_OpenSslHmac_HMAC_1Final
     (JNIEnv *env, jclass clazz, jlong ctx, jbyteArray md, jint mdOff,
         jint mdLen)
 {
@@ -129,7 +129,7 @@ Java_org_jitsi_srtp_OpenSslHmac_HMAC_1Final
  * Signature: (J[BIJJ)Z
  */
 JNIEXPORT jboolean JNICALL
-Java_org_jitsi_srtp_OpenSslHmac_HMAC_1Init_1ex
+Java_org_jitsi_srtp_crypto_OpenSslHmac_HMAC_1Init_1ex
     (JNIEnv *env, jclass clazz, jlong ctx, jbyteArray key, jint keyLen,
         jlong md, jlong impl)
 {
@@ -167,7 +167,7 @@ Java_org_jitsi_srtp_OpenSslHmac_HMAC_1Init_1ex
  * Signature: (J[BII)Z
  */
 JNIEXPORT jboolean JNICALL
-Java_org_jitsi_srtp_OpenSslHmac_HMAC_1Update
+Java_org_jitsi_srtp_crypto_OpenSslHmac_HMAC_1Update
     (JNIEnv *env, jclass clazz, jlong ctx, jbyteArray data, jint off, jint len)
 {
     jbyte *data_ = (*env)->GetPrimitiveArrayCritical(env, data, NULL);
