+++ src/main/native/SrtpCipherCtrOpenSsl.c
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "org_jitsi_srtp_SrtpCipherCtrOpenSsl.h"
+#include "org_jitsi_srtp_crypto_SrtpCipherCtrOpenSsl.h"
 
 #include <openssl/evp.h>
 #include <stdint.h>
@@ -26,7 +26,7 @@
  * Signature: ()J
  */
 JNIEXPORT jlong JNICALL
-Java_org_jitsi_srtp_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1create
+Java_org_jitsi_srtp_crypto_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1create
   (JNIEnv *env, jclass clazz)
 {
     EVP_CIPHER_CTX *ctx = EVP_CIPHER_CTX_new();
@@ -40,7 +40,7 @@ Java_org_jitsi_srtp_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1create
  * Signature: (J)V
  */
 JNIEXPORT void JNICALL
-Java_org_jitsi_srtp_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1destroy
+Java_org_jitsi_srtp_crypto_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1destroy
   (JNIEnv *env, jclass clazz, jlong ctx)
 {
     if (ctx) {
@@ -55,7 +55,7 @@ Java_org_jitsi_srtp_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1destroy
  * Signature: (J[B)Z
  */
 JNIEXPORT jboolean JNICALL
-Java_org_jitsi_srtp_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1init
+Java_org_jitsi_srtp_crypto_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1init
   (JNIEnv *env, jclass clazz, jlong ctx, jbyteArray key)
 {
     unsigned char key_[16];
@@ -69,7 +69,7 @@ Java_org_jitsi_srtp_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1init
  * Signature: (J[B[BII)Z
  */
 JNIEXPORT jboolean JNICALL
-Java_org_jitsi_srtp_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1process
+Java_org_jitsi_srtp_crypto_SrtpCipherCtrOpenSsl_AES128CTR_1CTX_1process
   (JNIEnv *env, jclass clazz, jlong ctx, jbyteArray iv, jbyteArray inOut, jint offset, jint len)
 {
     int ok = 0;
